pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orly-mcp"
version = "0.1.0"
description = "An MCP server for the ORLY tool."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # Or your preferred license
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp",
    "Flask>=2.0.0",
    "Jinja2>=3.0.0",
    "MarkupSafe>=2.0.0",
    "Werkzeug>=2.0.0",
    "itsdangerous>=2.0.0",
    "requests>=2.28.0",
    "Pillow>=11",
    "FontTools>=4.0.0",
]

[project.scripts]
orly-mcp = "orly_mcp.server:main"

[project.urls]
"Homepage" = "https://github.com/example/orly-mcp" # Replace with your actual URL
"Bug Tracker" = "https://github.com/example/orly-mcp/issues" # Replace with your actual URL
